refactor: 패키지 재배치 및 리팩터링
