+++ src/test/kotlin/blackjack/domain/card/CardsTest.kt
@@ -1,6 +1,5 @@
 package blackjack.domain.card
 
-import blackjack.domain.score.Score
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.collections.shouldContainExactly
 import io.kotest.matchers.shouldBe
